// hou2ed/app/src/navigation/types.ts
import { NativeStackNavigationProp } from "@react-navigation/native-stack";
import { RouteProp } from "@react-navigation/native";

export type RootStackParamList = {
  Splash: undefined;
  OnboardingScreen: undefined;
  RoleSelection: undefined;
  SignUp: undefined;
  Login: undefined;
  VerifyCode: { email?: string };
  ForgotPassword: undefined;
  Tabs: undefined;
  ListingDetails: { listingId?: string };
  ProviderListingDetails: { listingId: string };
  ApplyWizard: { listingId?: string };
  Thread: {
    threadId?: string;
    messageId: string;
    applicationId: string;
    propertyTitle: string;
    senderName: string;
    participantId?: string;
  };
  DevMenu: undefined;
  ProviderDashboard: undefined;
  AddListing: undefined;
  EditListing: {
    listingId: string;
    listingData?: {
      id: string;
      title: string;
      address: string;
      totalBeds: number;
      availableBeds: number;
      lastUpdated: string;
    };
  };
  AvailabilityUpdater: undefined;
  ListingWizard: undefined;
};

export type RootStackNavigationProp =
  NativeStackNavigationProp<RootStackParamList>;
export type RootStackRouteProp<T extends keyof RootStackParamList> = RouteProp<
  RootStackParamList,
  T
>;